style(db): rename file
