+++ backend/matcha/services/profile.py
@@ -15,7 +15,7 @@ from matcha.db.visit import db_put_visit, db_get_visit
 
 from matcha.db.last_connection import db_get_last_connection
 
-from matcha.db.db import (
+from matcha.db.user import (
     db_get_interests,
     db_set_user_profile_data,
     db_get_user_per_id,
